+++ test/ibmq/test_ibmq_jobmanager.py
@@ -28,7 +28,6 @@ from qiskit.providers.ibmq.managed.managedresults import ManagedResults
 from qiskit.providers.ibmq.managed.exceptions import (
     IBMQJobManagerJobNotFound, IBMQManagedResultDataNotAvailable, IBMQJobManagerInvalidStateError)
 from qiskit.providers.jobstatus import JobStatus, JOB_FINAL_STATES
-from qiskit.providers import JobError
 from qiskit.providers.ibmq.ibmqbackend import IBMQBackend
 from qiskit.providers.ibmq.exceptions import IBMQBackendError
 from qiskit.compiler import transpile, assemble
@@ -36,6 +35,7 @@ from qiskit.compiler import transpile, assemble
 from ..ibmqtestcase import IBMQTestCase
 from ..decorators import requires_provider
 from ..fake_account_client import BaseFakeAccountClient
+from ..utils import cancel_job
 
 
 class TestIBMQJobManager(IBMQTestCase):
@@ -373,15 +373,7 @@ class TestResultManager(IBMQTestCase):
         job_set = self._jm.run(circs, backend=backend)
         jobs = job_set.jobs()
         cjob = jobs[1]
-        cancelled = False
-        for _ in range(2):
-            # Try twice in case job is not in a cancellable state
-            try:
-                cancelled = cjob.cancel()
-                if cancelled:
-                    break
-            except JobError:
-                pass
+        cancelled = cancel_job(cjob)
 
         result_manager = job_set.results()
         if cancelled: